feat: add UserIdentifier entity, modify listAllIds, rename ListAllUserIdsUseCase
