+++ src/CompositionRoot.ts
@@ -12,7 +12,7 @@ import { GetUsersByIdsUseCase } from "./domain/usecases/GetUsersByIdsUseCase";
 import { ListMetadataUseCase } from "./domain/usecases/ListMetadataUseCase";
 import { ListUsersUseCase } from "./domain/usecases/ListUsersUseCase";
 import { ListAllUsersUseCase } from "./domain/usecases/ListAllUsersUseCase";
-import { ListAllUserIdsUseCase } from "./domain/usecases/ListAllUserIdsUseCase";
+import { ListAllUserIdentifiersUseCase } from "./domain/usecases/ListAllUserIdentifiersUseCase";
 import { RemoveUsersUseCase } from "./domain/usecases/RemoveUsersUseCase";
 import { SaveColumnsUseCase } from "./domain/usecases/SaveColumnsUseCase";
 import { SaveUserOrgUnitUseCase } from "./domain/usecases/SaveUserOrgUnitUseCase";
@@ -51,7 +51,7 @@ export function getCompositionRoot(instance: Instance) {
             getCurrent: new GetCurrentUserUseCase(userRepository),
             list: new ListUsersUseCase(userRepository),
             listAll: new ListAllUsersUseCase(userRepository),
-            listAllIds: new ListAllUserIdsUseCase(userRepository),
+            listAllIdentifiers: new ListAllUserIdentifiersUseCase(userRepository),
             get: new GetUsersByIdsUseCase(userRepository),
             save: new SaveUsersUseCase(userRepository),
             saveStatus: new SaveUserStatusUseCase(userRepository),
